Add asynchronous email and emails from templates
The template emails take templates with markdown formatting.  First, the
template is rendered.  Then, the markdown is converted into HTML to be
the HTML message, while the raw markdown is the plain-text message.  The
current system is built around Django's send_mail.  We will probably need
to add support for, or switch entirely to, send_mass_email.  We also may
need to swap over to the EmailMessage class directly to set additional
headers, like Reply-To.

For testing purposes, we want the task queue to run synchronously, so we can
examine the emails that get produced.  Overriding the settings on the test
cases doesn't work (probably too late for Django-Q to notice), so in
settings.py, we guess whether we're in testing, and set the sync setting
appropriately.  We also increase the logging threshold in that case, to
avoid log spam from Django-Q.