Separate email templating from email sending
Many (all?) emails will be sent to Members, and their contents will be
member-specific.  Instead of inventing an intermediate form to pass
this data from Members into the email sending, this lets us format
emails directly from Members into EmailMultiAlternatives instances.
This will be necessary anyway to use features like Reply-To, so we
might as well start using it here.

The proposed flow to send a message to a number of Members would be to
use a list comprehension to prepare an email to each of the members,
and then pass that list of emails to lib.email.send_messages_async.