Send email when a gig is created
We listen for the post_save signal from Gigs, and send an email to
all associations of that band.  Much of the display logic is pushed
into the template; this allows translation and localization to happen
based on the language of the user to whom the email is being sent.

The get_confirm_urls in gig/helpers.py is just a placeholder; we need
the relevant endpoints to be built before we can hook those up.

Much of this code should be reusable for the update and reminder
emails, but so far there's no generalization of the relevant code.