Include change information in the gig edit email
This involves some tricky formatting, which is easier to set up in Python,
rather than the template.  But this requires us to ensure that translations
aren't resolved until we're in the correct language context.  This is
accomplished with the lazy string join and the use of functions that
will format dates when called.

The HistoricalRecords of the gigs only have the attributes of the gigs,
not any of the methods.  Therefore, we turn status_string from a property
into a method, so we can call it on the Gig class with values coming from
a Historical Record.