Separate email sending from member classes
This introduces a EmailRecipient class to be used to pass information about
an email recipient into prepare_email.  This lets us move prepare_email
into the email library, and means we don't have to work to make Member and
Invite look like each other.  We can't pull as much information about these
objects out in the template, but arguably it's better to be more explicit
about the attributes needed when calling the template.