Set language via login page
This hooks up a view provided by Django to switch language, but the requires a POST.  We
jump though a bunch of hoops to make a bunch of form submit buttons looks like regular
links.  It would have probably been easier to just write the endpoint to switch the
language ourselves.

It was hard to get the local language names displaying correctly.  The language_name_local
filter gives us a localized name, but it discards to sublanguage part.  That is, both
'en-us' and 'en-uk' give 'English'.  The language description in LANGUAGES does have the
sublanguage description, but it is in the current language, not the local language for
each.  The current code hacks this by using 'language_name_local', but then appending the
sublanguage code, if it exists.  This works well for US and UK; it's probably not a good
general solution, though.