Set the user's language from the invitation if appropriate
Before, we just translated the accept page with this language.  But this meant that the user sign-up flow
would not be properly translated.  We can set a user's language before they log in, so we'll do that when
they hit the invite page.  That way, all subsequent pages will be in the correct language.

We will only set the language if they are not logged in and they don't already have a lanaguage cookie
set.  The former case should never be true without the latter, but we might was well be safe here.

Note that we have to set the current language before rendering a response, but we can only set the
language cookie after we have a response that's been rendered.  Most annoying.  This is why the language
activation code is split between inline code and a function executed later.