Cleaner solution for assoc tfparams
The parameter and name are now send in the POST body, thanks to a
htmx extension.  This simplifies the URL structure and it means we
don't have to reload the checkbox to get a new URL.  It also lets
us submit multiple parameters in a single request, should that be
desired in the future.