Reduce workload on queued tasks
This investigation was started because invitations were being sent and resent (and resent...) every 5 minutes. My suspicion is that the model callback task was failing and getting re-enqueued. Looking at the previous implementation, it queued a task to queue a task. This PR simplifies that to just one task per invitation email.

While I was at it, I changed the Django-Q behavior to not attempt to retry failed tasks by default. This is definitely important for email so we don't accidentally spam people when things don't work right, and I think this default is a safe one in general.