Let enter key submit forms
I noticed that I was not able to type my username & password and then press enter to log in. This snippet of JS is the culprit. I'm sure it was added to fix a problem at some point, but in my testing I can't find any screen where pressing the enter key does something unexpected. I think we should allow default form behavior, and if an issue does come up, figure out a more targeted fix.