Cancelled is Canceled
One of the problems uncovered while implementing this was that the status code was GigStatusChoices.CANCELLED while the preference was named `hide_canceled_gigs`. We have to agree with ourselves to make this work.

According to [the internet](https://writer.com/blog/canceled-cancelled/) "Canceled" is the preferred speling for Americans. And since I am American, I am arbitrarily choosing this speling.