Initialize the calendar with a known timezone
Try to always show a relevant timezone, regardless of where the user's computer happens to be in the world.
Infer the timezone based on the current user's first band assocation.
If a user is in more than one band, and those bands are in different timezones, then any other timezones will be localized to the first band's timezone. Not sure right now how to do better than this.